ai/requirements.txt
torch
torchvision
streamlit
python-chess
Pillow
tqdm